MS-34 refactor(routing): code refactoring, removing unnecessary code, correcting variable names
