+++ src/pages/basket/basket.ts
@@ -1,8 +1,8 @@
 export function renderBasketContent() {
-  const basketText = "This is the Basket page!";
+  const basketTemporaryContent = "This is the Basket page!";
   return `
       <div class="basket-page">
-          ${basketText}
+          ${basketTemporaryContent}
       </div>
   `;
 }
